improved debug settings
