+++ CI-CD/Localizable.xcstrings
@@ -2033,6 +2033,10 @@
         }
       }
     },
+    "Xcode Cloud" : {
+      "comment" : "A section in the debug settings for managing App Store Connect credentials.",
+      "isCommentAutoGenerated" : true
+    },
     "You need an API key with the 'Admin' role from App Store Connect to start builds with external deployments. API keys with the 'Developer' role cannot be used for this" : {
       "localizations" : {
         "nl" : {
